+++ periodic-ingester/Program.fs
@@ -1,4 +1,4 @@
-﻿module FeedConsumerTemplate.Program
+﻿module PeriodicIngesterTemplate.Ingester.Program
 
 open Serilog
 open System
@@ -16,49 +16,62 @@ type Configuration(tryGet) =
     member _.CosmosDatabase =               get "EQUINOX_COSMOS_DATABASE"
     member _.CosmosContainer =              get "EQUINOX_COSMOS_CONTAINER"
 
+    member _.PrometheusPort =               tryGet "PROMETHEUS_PORT" |> Option.map int
+
     member _.BaseUri =                      get "API_BASE_URI"
     member _.Group =                        get "API_CONSUMER_GROUP"
 
 module Args =
 
     open Argu
-    [<NoEquality; NoComparison>]
-    type Parameters =
-        | [<AltCommandLine "-V"; Unique>]   Verbose
 
-        | [<AltCommandLine "-g"; Unique>]   Group of string
-        | [<AltCommandLine "-f"; Unique>]   BaseUri of string
+    type [<NoEquality; NoComparison>] Parameters =
+        | [<AltCommandLine "-V"; Unique>]   Verbose
+        | [<AltCommandLine "-p"; Unique>]   PrometheusPort of int
 
         | [<AltCommandLine "-r"; Unique>]   MaxReadAhead of int
-        | [<AltCommandLine "-w"; Unique>]   FcsDop of int
         | [<AltCommandLine "-t"; Unique>]   TicketsDop of int
 
-        | [<CliPrefix(CliPrefix.None); Unique; Last>] Cosmos of ParseResults<CosmosParameters>
+        | [<CliPrefix(CliPrefix.None); Unique; Last>] Feed of ParseResults<FeedParameters>
         interface IArgParserTemplate with
             member a.Usage = a |> function
                 | Verbose _ ->              "request verbose logging."
-                | Group _ ->                "specify Api Consumer Group Id. (optional if environment variable API_CONSUMER_GROUP specified)"
-                | BaseUri _ ->              "specify Api endpoint. (optional if environment variable API_BASE_URI specified)"
+                | PrometheusPort _ ->       "port from which to expose a Prometheus /metrics endpoint. Default: off (optional if environment variable PROMETHEUS_PORT specified)"
                 | MaxReadAhead _ ->         "maximum number of batches to let processing get ahead of completion. Default: 8."
-                | FcsDop _ ->               "maximum number of FCs to process in parallel. Default: 4"
-                | TicketsDop _ ->           "maximum number of Tickets to process in parallel (per FC). Default: 4"
-                | Cosmos _ ->               "Cosmos Store parameters."
+                | TicketsDop _ ->           "maximum number of Tickets to process in parallel. Default: 4"
+                | Feed _ ->                 "Feed parameters."
     and Arguments(c : Configuration, a : ParseResults<Parameters>) =
         member val Verbose =                a.Contains Parameters.Verbose
-        member val Group =                  a.TryGetResult Group        |> Option.defaultWith (fun () -> c.Group)
-        member val BaseUri =                a.TryGetResult BaseUri      |> Option.defaultWith (fun () -> c.BaseUri) |> Uri
-        member val MaxReadAhead =           a.GetResult(MaxReadAhead,8)
-        member val FcsDop =                 a.TryGetResult FcsDop       |> Option.defaultValue 4
-        member val TicketsDop =             a.TryGetResult TicketsDop   |> Option.defaultValue 4
+        member val PrometheusPort =         a.TryGetResult PrometheusPort |> Option.orElseWith (fun () -> c.PrometheusPort)
+        member val MaxReadAhead =           a.GetResult(MaxReadAhead, 8)
+        member val TicketsDop =             a.GetResult(TicketsDop, 4)
         member val StatsInterval =          TimeSpan.FromMinutes 1.
         member val StateInterval =          TimeSpan.FromMinutes 5.
         member val CheckpointInterval =     TimeSpan.FromHours 1.
+        member val Feed : FeedArguments =
+            match a.TryGetSubCommand() with
+            | Some (Feed feed) -> FeedArguments(c, feed)
+            | _ -> raise (MissingArg "Must specify feed")
+    and [<NoEquality; NoComparison>] FeedParameters =
+        | [<AltCommandLine "-g"; Unique>]   Group of string
+        | [<AltCommandLine "-f"; Unique>]   BaseUri of string
+        | [<CliPrefix(CliPrefix.None); Unique; Last>] Cosmos of ParseResults<CosmosParameters>
+        interface IArgParserTemplate with
+            member a.Usage = a |> function
+                | Group _ ->                "specify Api Consumer Group Id. (optional if environment variable API_CONSUMER_GROUP specified)"
+                | BaseUri _ ->              "specify Api endpoint. (optional if environment variable API_BASE_URI specified)"
+                | Cosmos _ ->               "Cosmos Store parameters."
+    and FeedArguments(c : Configuration, a : ParseResults<FeedParameters>) =
+        member val SourceId =               a.TryGetResult Group   |> Option.defaultWith (fun () -> c.Group)   |> Propulsion.Feed.SourceId.parse
+        member val BaseUri =                a.TryGetResult BaseUri |> Option.defaultWith (fun () -> c.BaseUri) |> Uri
+        member val RefreshInterval =        TimeSpan.FromHours 1.
         member val Cosmos : CosmosArguments =
             match a.TryGetSubCommand() with
             | Some (Cosmos cosmos) -> CosmosArguments(c, cosmos)
             | _ -> raise (MissingArg "Must specify cosmos")
     and [<NoEquality; NoComparison>] CosmosParameters =
-        | [<AltCommandLine "-cm">]          ConnectionMode of Microsoft.Azure.Cosmos.ConnectionMode
+        | [<AltCommandLine "-V"; Unique>]   Verbose
+        | [<AltCommandLine "-m">]          ConnectionMode of Microsoft.Azure.Cosmos.ConnectionMode
         | [<AltCommandLine "-s">]           Connection of string
         | [<AltCommandLine "-d">]           Database of string
         | [<AltCommandLine "-c">]           Container of string
@@ -67,6 +80,7 @@ module Args =
         | [<AltCommandLine "-rt">]          RetriesWaitTime of float
         interface IArgParserTemplate with
             member a.Usage = a |> function
+                | Verbose _ ->              "request verbose logging."
                 | ConnectionMode _ ->       "override the connection mode. Default: Direct."
                 | Connection _ ->           "specify a connection string for a Cosmos account. (optional if environment variable EQUINOX_COSMOS_CONNECTION specified)"
                 | Database _ ->             "specify a database name for Cosmos store. (optional if environment variable EQUINOX_COSMOS_DATABASE specified)"
@@ -75,14 +89,15 @@ module Args =
                 | Retries _ ->              "specify operation retries (default: 9)."
                 | RetriesWaitTime _ ->      "specify max wait-time for retry when being throttled by Cosmos in seconds (default: 30)"
     and CosmosArguments(c : Configuration, a : ParseResults<CosmosParameters>) =
-        let discovery =                     a.TryGetResult Connection   |> Option.defaultWith (fun () -> c.CosmosConnection) |> Equinox.CosmosStore.Discovery.ConnectionString
+        let discovery =                     a.TryGetResult Connection |> Option.defaultWith (fun () -> c.CosmosConnection) |> Equinox.CosmosStore.Discovery.ConnectionString
         let mode =                          a.TryGetResult ConnectionMode
-        let timeout =                       a.GetResult(Timeout, 30.)   |> TimeSpan.FromSeconds
+        let timeout =                       a.GetResult(Timeout, 30.) |> TimeSpan.FromSeconds
         let retries =                       a.GetResult(Retries, 9)
         let maxRetryWaitTime =              a.GetResult(RetriesWaitTime, 30.) |> TimeSpan.FromSeconds
         let connector =                     Equinox.CosmosStore.CosmosStoreConnector(discovery, timeout, retries, maxRetryWaitTime, ?mode=mode)
-        let database =                      a.TryGetResult Database     |> Option.defaultWith (fun () -> c.CosmosDatabase)
-        let container =                     a.TryGetResult Container    |> Option.defaultWith (fun () -> c.CosmosContainer)
+        let database =                      a.TryGetResult Database   |> Option.defaultWith (fun () -> c.CosmosDatabase)
+        let container =                     a.TryGetResult Container  |> Option.defaultWith (fun () -> c.CosmosContainer)
+        member val Verbose =                a.Contains Verbose
         member _.Connect() =                connector.ConnectStore("Main", database, container)
 
     /// Parse the commandline; can throw exceptions in response to missing arguments and/or `-h`/`--help` args
@@ -91,58 +106,47 @@ module Args =
         let parser = ArgumentParser.Create<Parameters>(programName = programName)
         Arguments(Configuration tryGetConfigValue, parser.ParseCommandLine argv)
 
-let [<Literal>] AppName = "FeedConsumerTemplate"
+let [<Literal>] AppName = "PeriodicIngesterTemplate"
 
 let build (args : Args.Arguments) =
     let cache = Equinox.Cache (AppName, sizeMb = 10)
-    let context = args.Cosmos.Connect() |> Async.RunSynchronously |> CosmosStoreContext.create
+    let feed = args.Feed
+    let context = feed.Cosmos.Connect() |> Async.RunSynchronously |> CosmosStoreContext.create
 
     let sink =
-        let handle = Ingester.handle args.TicketsDop
         let stats = Ingester.Stats(Log.Logger, args.StatsInterval, args.StateInterval)
-        Propulsion.Streams.StreamsProjector.Start(Log.Logger, args.MaxReadAhead, args.FcsDop, handle, stats, args.StatsInterval)
+        Propulsion.Streams.StreamsProjector.Start(Log.Logger, args.MaxReadAhead, args.TicketsDop, Ingester.handle, stats, args.StatsInterval)
     let pumpSource =
-        let sourceId, tailSleepInterval = Propulsion.Feed.SourceId.parse args.Group, TimeSpan.FromSeconds 1.
-        let checkpoints = Propulsion.Feed.ReaderCheckpoint.CosmosStore.create Equinox.log (context, cache)
-        let feed = ApiClient.TicketsFeed args.BaseUri
+        let checkpoints = Propulsion.Feed.ReaderCheckpoint.CosmosStore.create Log.Logger (context, cache)
+        let client = ApiClient.TicketsFeed feed.BaseUri
         let source =
-            Propulsion.Feed.FeedSource(
-                Log.Logger, args.StatsInterval, sourceId, tailSleepInterval,
-                checkpoints, args.CheckpointInterval, feed.Poll, sink)
-        source.Pump feed.ReadTranches
+            Propulsion.Feed.PeriodicSource(
+                Log.Logger, args.StatsInterval, feed.SourceId,
+                checkpoints, args.CheckpointInterval,
+                client.Crawl, feed.RefreshInterval,
+                sink)
+        source.Pump()
     sink, pumpSource
 
-[<AutoOpen>]
-module ConsumerExt =
-    type Propulsion.Pipeline with
-        member sink.AwaitWithStopOnCancellation() = async {
-            let! ct = Async.CancellationToken
-            use _ = ct.Register(fun () -> sink.Stop())
-            return! sink.AwaitCompletion()
-        }
+// A typical app will likely have health checks etc, implying the wireup would be via `endpoints.MapMetrics()` and thus not use this ugly code directly
+let startMetricsServer port : IDisposable =
+    let metricsServer = new Prometheus.KestrelMetricServer(port = port)
+    let ms = metricsServer.Start()
+    Log.Information("Prometheus /metrics endpoint on port {port}", port)
+    { new IDisposable with member x.Dispose() = ms.Stop(); (metricsServer :> IDisposable).Dispose() }
 
 let run args = async {
     let sink, pumpSource = build args
+    use _metricsServer : IDisposable = args.PrometheusPort |> Option.map startMetricsServer |> Option.toObj
     do! Async.Parallel [ pumpSource; sink.AwaitWithStopOnCancellation() ] |> Async.Ignore<unit[]>
     return if sink.RanToCompletion then 0 else 3
 }
 
-[<System.Runtime.CompilerServices.Extension>]
-type Logging() =
-
-    [<System.Runtime.CompilerServices.Extension>]
-    static member Configure(configuration : LoggerConfiguration, ?verbose) =
-        configuration
-            .Destructure.FSharpTypes()
-            .Enrich.FromLogContext()
-        |> fun c -> if verbose = Some true then c.MinimumLevel.Debug() else c
-        |> fun c -> let theme = Sinks.SystemConsole.Themes.AnsiConsoleTheme.Code
-                    c.WriteTo.Console(theme=theme, outputTemplate="[{Timestamp:HH:mm:ss} {Level:u3}] {Message:lj} {NewLine}{Exception}")
-
 [<EntryPoint>]
 let main argv =
     try let args = Args.parse EnvVar.tryGet argv
-        try Log.Logger <- LoggerConfiguration().Configure(verbose=args.Verbose).CreateLogger()
+        try let metrics = Sinks.equinoxAndPropulsionFeedConsumerMetrics (Sinks.tags AppName) args.Feed.SourceId
+            Log.Logger <- LoggerConfiguration().Configure(args.Verbose).Sinks(metrics, args.Feed.Cosmos.Verbose).CreateLogger()
             try run args |> Async.RunSynchronously
             with e when not (e :? MissingArg) -> Log.Fatal(e, "Exiting"); 2
         finally Log.CloseAndFlush()